cli: optimize command execution logic
1. move permission check into command executor
2. code optimization

+++ cli/src/cmd/reboot.rs
@@ -1,39 +1,39 @@
 use log::debug;
 use common::util::fs;
 
+use crate::cli::SyscareCLI;
 use crate::boot::{BootManager, RebootOption};
 
-use super::{CommandExecutor, CommandArguments};
+use super::{CommandArguments, CommandExecutor};
 
 pub struct RebootCommandExecutor;
 
 impl CommandExecutor for RebootCommandExecutor {
     fn invoke(&self, args: &CommandArguments) -> std::io::Result<i32> {
-        match args {
-            CommandArguments::RebootArguments(target, force) => {
-                if !force {
-                    debug!("Syncing filesystem");
-                    fs::sync();
+        SyscareCLI::check_root_permission()?;
+
+        if let CommandArguments::RebootArguments(target, force) = args {
+            if !force {
+                debug!("Syncing filesystem");
+                fs::sync();
+            }
+
+            debug!("Preparing for reboot");
+            BootManager::load_kernel(target.as_ref())?;
+
+
+            BootManager::reboot(match force {
+                false => {
+                    debug!("Rebooting system");
+                    RebootOption::Normal
+                },
+                true => {
+                    debug!("Force rebooting system");
+                    RebootOption::Forced
                 }
-
-                debug!("Preparing for reboot");
-                BootManager::load_kernel(target.as_ref())?;
-
-
-                BootManager::reboot(match force {
-                    false => {
-                        debug!("Rebooting system");
-                        RebootOption::Normal
-                    },
-                    true => {
-                        debug!("Force rebooting system");
-                        RebootOption::Forced
-                    }
-                })?;
-
-                Ok(0)
-            },
-            _ => unreachable!(),
+            })?;
         }
+
+        Ok(0)
     }
 }
